Adding the GET method

+++ RESTAPI/Models/Course.cs
@@ -6,7 +6,7 @@ namespace RESTAPI.Models
         public string CourseName { get; set; }
         public int CourseNumber { get; set; }
         public string CourseDepartment { get; set; }
-        public int CourseCredits { get; set; }
+        public double CourseCredits { get; set; }
 
         public string CourseInstructor { get; set; }
     }
